Take notary private key and nonce generation from hosts
